refactor(tasks.py): Update BTE calls
