+++ cl/scrapers/tasks.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import print_function
-
+from cl.scrapers.docker_helpers import get_page_count, document_extract, \
+    convert_audio
 import os
 import logging
 import random
@@ -38,79 +39,6 @@ from cl.search.tasks import add_items_to_solr
 from juriscraper.pacer import PacerSession, CaseQuery
 
 logger = logging.getLogger(__name__)
-
-
-def get_page_count(path, extension=None):
-    """Helper method to call get_page_count.
-
-    :param path: Location of file
-    :param extension: File extension
-    :type: str
-    :return: dictionary containing page count.
-    """
-    with open(path, "rb") as file:
-        f = file.read()
-    try:
-        return requests.post(
-            "http://cl-binary-transformers-and-extractors:80/get_page_count",
-            files={"file": (os.path.basename(path), f)},
-            timeout=300,
-        ).json()["pg_count"]
-    except Timeout:
-        return {
-            "err": Timeout,
-            "msg": "Timeout error occurred; Page count failed.",
-        }
-    except:
-        return {"err": "An unknown error occured."}
-
-
-def process_doc(path, do_ocr=False):
-    """Helper method to call task extract doc content
-
-    :param path:
-    :param do_ocr:
-    :return:
-    """
-    with open(path, "rb") as file:
-        f = file.read()
-    try:
-        return requests.post(
-            "http://cl-binary-transformers-and-extractors:80/extract_doc_content",
-            files={"file": (os.path.basename(path), f)},
-            params={"do_ocr": do_ocr},
-            timeout=3600,
-        ).json()
-    except Timeout:
-        return {
-            "err": Timeout,
-            "msg": "Timeout error occurred; Failed conversion",
-        }
-    except:
-        return {"err": "An unknown error occured."}
-
-
-def send_file_to_convert_audio(filepath):
-    """This is a helper function to call convert_audio_file.
-
-    :param filepath:
-    :return:
-    """
-    with open(filepath, "rb") as file:
-        f = file.read()
-    try:
-        return requests.post(
-            "http://cl-binary-transformers-and-extractors:80/convert_audio_file",
-            files={"file": (os.path.basename(filepath), f)},
-            timeout=3600,
-        )
-    except Timeout:
-        return {
-            "err": Timeout,
-            "msg": "Timeout error occurred; Failed conversion",
-        }
-    except:
-        return {"err": "An unknown error occured."}
 
 
 def update_document_from_text(opinion):
@@ -166,11 +94,11 @@ def extract_doc_content(pk, do_ocr=False, citation_jitter=False):
     path = opinion.local_path.path
 
     extension = path.split(".")[-1]
-    response = process_doc(path, do_ocr=do_ocr)
+    response = document_extract(path, do_ocr)
     content = response["content"]
     success = response["err"]
     # Do page count, if possible
-    opinion.page_count = get_page_count(path)
+    opinion.page_count = get_page_count(path)['pg_count']
 
     # Do blocked status
     if extension in ["html", "wpd"]:
@@ -230,7 +158,7 @@ def extract_recap_pdf(pks, skip_ocr=False, check_if_needed=True):
             continue
 
         path = rd.filepath_local.path
-        response = process_doc(path, do_ocr=True)
+        response = document_extract(path, do_ocr=True)
         content = response["content"]
         err = response["err"]
 
@@ -342,28 +270,37 @@ def process_audio_file(pk):
     """
     af = Audio.objects.get(pk=pk)
     tmp_path = os.path.join("/tmp", "audio_" + uuid.uuid4().hex + ".mp3")
-    r = send_file_to_convert_audio(af.local_path_original_file.path)
-    with open(tmp_path, "w") as mp3:
-        mp3.write(r.content)
-    set_mp3_meta_data(af, tmp_path)
-    try:
-        with open(tmp_path, "r") as mp3:
-            cf = ContentFile(mp3.read())
-            file_name = trunc(best_case_name(af).lower(), 72) + "_cl.mp3"
-            af.file_with_date = af.docket.date_argued
-            af.local_path_mp3.save(file_name, cf, save=False)
-    except:
-        msg = (
-            "Unable to save mp3 to audio_file in scraper.tasks."
-            "process_audio_file for item: %s\n"
-            "Traceback:\n"
-            "%s" % (af.pk, traceback.format_exc())
-        )
-        print(msg)
-        ErrorLog.objects.create(
-            log_level="CRITICAL", court=af.docket.court, message=msg
-        )
+    response = convert_audio(af.local_path_original_file.path)
 
+    file = response["content"]
+    err = response["err"]
+    if not err == Timeout:
+        with open(tmp_path, "w") as mp3:
+            mp3.write(file.content)
+        set_mp3_meta_data(af, tmp_path)
+        try:
+            with open(tmp_path, "r") as mp3:
+                cf = ContentFile(mp3.read())
+                file_name = trunc(best_case_name(af).lower(), 72) + "_cl.mp3"
+                af.file_with_date = af.docket.date_argued
+                af.local_path_mp3.save(file_name, cf, save=False)
+        except:
+            msg = (
+                "Unable to save mp3 to audio_file in scraper.tasks."
+                "process_audio_file for item: %s\n"
+                "Traceback:\n"
+                "%s" % (af.pk, traceback.format_exc())
+            )
+            print(msg)
+            ErrorLog.objects.create(
+                log_level="CRITICAL", court=af.docket.court, message=msg
+            )
+    else:
+        ErrorLog.objects.create(
+            log_level="CRITICAL",
+            court=af.docket.court,
+            message="Timeout occurred in docker container for %s" % (af.pk),
+        )
     af.duration = eyed3.load(tmp_path).info.time_secs
     af.processing_complete = True
     af.save()
