+++ cl/search/management/commands/cl_count_recap_pages.py
@@ -1,8 +1,7 @@
 from collections import Counter
 
-
 from cl.lib.command_utils import VerboseCommand, logger
-from cl.scrapers.tasks import get_page_count
+from cl.scrapers.docker_helpers import get_page_count
 from cl.search.models import RECAPDocument
 
 
@@ -24,7 +23,7 @@ class Command(VerboseCommand):
                 cnt["no_file"] += 1
             else:
                 extension = path.split(".")[-1]
-                count = get_page_count(path, extension)
+                count = get_page_count(path, extension)['pg_count']
                 rd.page_count = count
                 rd.save(do_extraction=False, index=False)
                 cnt["successes"] += 1
